suffix_array수정
