+++ suffix_array/suffix_array_library.py
@@ -1,30 +1,14 @@
 import time
-import functools
-import resource
-import platform
-
+import pydivsufsort
 
 def build_suffix_array(text: str) -> list:
-    """
-    슬라이스 복사 없이 비교하는 메모리 절약형 suffix array 생성
-    """
-    def compare(i, j):
-        while i < len(text) and j < len(text):
-            if text[i] != text[j]:
-                return -1 if text[i] < text[j] else 1
-            i += 1
-            j += 1
-        return -1 if i == len(text) and j != len(text) else (1 if j == len(text) and i != len(text) else 0)
-
-    return sorted(range(len(text)), key=functools.cmp_to_key(compare))
-
+    return list(pydivsufsort.divsufsort(text.encode()))
 
 def _compare_prefix(text: str, pos: int, pattern: str) -> int:
     slice_ref = text[pos : pos + len(pattern)]
     if slice_ref == pattern:
         return 0
     return -1 if slice_ref < pattern else 1
-
 
 def find_exact_matches(sa: list, text: str, pattern: str) -> list:
     n = len(sa)
@@ -87,40 +71,25 @@ def align_single_read(sa: list, reference: str, read: str, k: int) -> int:
     return best_pos if best_mm <= k else -1
 
 
-def align_all_reads_blockwise(reference: str, reads: list, k: int, block_size: int = 500_000, overlap: int = 100) -> dict:
-    """
-    reference를 block_size 단위로 나누고, 블록 간 overlap을 추가하여 suffix array 생성 및 매칭
-    """
+def align_all_reads(reference: str, reads: list, k: int) -> dict:
+    print("🔄 Suffix array 생성 중… (C 라이브러리 divsufsort 사용)")
+    t0 = time.time()
+    sa = build_suffix_array(reference)
+    t1 = time.time()
+    print(f"✅ SA 생성 완료 (소요: {t1-t0:.2f}초, length={len(sa)})\n")
+
     total_reads = len(reads)
     alignments = {}
-    matched = [False] * total_reads
+    PRINT_INTERVAL = max(1, total_reads // 100)
 
-    num_blocks = (len(reference) + block_size - 1) // block_size
+    for idx, read in enumerate(reads):
+        pos = align_single_read(sa, reference, read, k)
+        alignments[idx] = pos
+        if (idx + 1) % PRINT_INTERVAL == 0 or (idx + 1) == total_reads:
+            pct = (idx + 1) / total_reads * 100
+            print(f"🔄 처리 중: {idx + 1}/{total_reads} reads ({pct:.1f}%) 완료")
 
-    for block_idx in range(num_blocks):
-        block_start = block_idx * block_size
-        block_end = min((block_idx + 1) * block_size + overlap, len(reference))  # ← overlap 추가
-        block_ref = reference[block_start:block_end]
-
-        print(f"🔄 Block {block_idx + 1}/{num_blocks} (위치 {block_start}~{block_end}) 처리 중...")
-        sa = build_suffix_array(block_ref)
-
-        for idx, read in enumerate(reads):
-            if matched[idx]:
-                continue
-
-            pos_in_block = align_single_read(sa, block_ref, read, k)
-            if pos_in_block >= 0:
-                global_pos = block_start + pos_in_block
-                if global_pos + len(read) <= len(reference):  # 전체 범위 초과 방지
-                    alignments[idx] = global_pos
-                    matched[idx] = True
-
-    # 매칭 실패한 read는 -1로 저장
-    for idx in range(total_reads):
-        if not matched[idx]:
-            alignments[idx] = -1
-
+    print()
     return alignments
 
 
@@ -135,29 +104,26 @@ def reconstruct_mammoth(reference: str, reads: list, alignments: dict) -> str:
         if pos < 0:
             continue
         for i, base in enumerate(read):
-            if pos + i < len(ref_list):
-                ref_list[pos + i] = base
+            ref_list[pos + i] = base
 
     return "".join(ref_list)
 
 
 def main():
     k = 2  # 허용 mismatch 개수
-    BLOCK_SIZE = 500_000  # reference 블록 크기
-    OVERLAP = 100  # 블록 간 오버랩
 
     start_time = time.time()
 
-    with open("../genome_generation/3_3_reference_100M.txt", "r") as f_ref:
+    with open("../genome_generation/4_reference_1B.txt", "r") as f_ref:
         reference = f_ref.read().strip()
 
-    with open("../genome_generation/3_3_mammoth_reads_100K.txt", "r") as f_reads:
+    with open("../genome_generation/4_mammoth_reads_10M.txt", "r") as f_reads:
         reads = [line.strip() for line in f_reads if line.strip()]
 
-    with open("../genome_generation/3_3_ground_truth_100K.txt", "r") as f_gt:
+    with open("../genome_generation/4_ground_truth_10M.txt", "r") as f_gt:
         ground_truth = [int(line.strip()) for line in f_gt if line.strip()]
 
-    alignments = align_all_reads_blockwise(reference, reads, k, block_size=BLOCK_SIZE, overlap=OVERLAP)
+    alignments = align_all_reads(reference, reads, k)
 
     total_reads = len(reads)
     correct_count = 0
@@ -178,11 +144,6 @@ def main():
 
     end_time = time.time()
     print(f"⏱ 전체 실행 시간: {end_time - start_time:.2f}초")
-
-    # ✅ 최대 메모리 사용량 출력 (macOS 기준)
-    usage = resource.getrusage(resource.RUSAGE_SELF)
-    memory_mb = usage.ru_maxrss / (1024 * 1024)  # macOS는 바이트 단위
-    print(f"💾 최대 메모리 사용량: {memory_mb:.2f} MB")
 
 
 if __name__ == "__main__":
